Refactoring of site statistics to new file.  Add to admin stats page along with a regnerate button
