+++ caper/templates/pages/index.html
@@ -358,13 +358,13 @@
                 <tbody>
                     <tr>
                         <td>Projects</td>
-                        <td>{{ repo_stats.public_proj_count }}</td>
-                        <td>{{ repo_stats.all_private_proj_count }}</td>
+                        <td>{{ site_stats.public_proj_count }}</td>
+                        <td>{{ site_stats.all_private_proj_count }}</td>
                     </tr>
                     <tr>
                         <td>Samples</td>
-                        <td>{{ repo_stats.public_sample_count }}</td>
-                        <td>{{ repo_stats.all_private_sample_count }}</td>
+                        <td>{{ site_stats.public_sample_count }}</td>
+                        <td>{{ site_stats.all_private_sample_count }}</td>
                     </tr>
                 </tbody>
             </table>
